requirements.txt
# Core dependencies
pyyaml>=6.0,<7.0
termcolor>=2.3.0,<4.0
rich>=13.0.0,<15.0
readchar>=4.0.0,<5.0

# Performance dependencies (optional)
ijson>=3.2.0,<4.0  # Streaming JSON parser for large files

# Test dependencies  
pytest>=8.0.0,<9.0
pytest-cov>=4.0.0,<7.0
pytest-asyncio>=0.24.0,<1.0

# Security scanning
safety>=3.0.0,<4.0

# Linting and formatting
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
pylint>=2.17.0
pre-commit>=3.0.0